+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to HaloCommunityBot will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-05-12
+
+### Added
+
+* support @handle format and per-channel keyword filtering for YouTube channels
+
+## [1.0.7] - 2026-05-12
+
+### Added
+
+* add per-channel keyword filtering for youtube videos
+
 ## [1.0.6] - 2026-05-12
 
 ### Changed